Organize validator implementation package
