+++ every_eval_ever/validator/cli.py
@@ -1,0 +1,9 @@
+"""Canonical module entry point for the validation CLI."""
+
+from .validate import main
+
+__all__ = ['main']
+
+
+if __name__ == '__main__':
+    raise SystemExit(main())
